names: unify PairedEdge/Comparison into Alignment evidence type
Replace rigour's two near-twin output types — PairedEdge from
pair_symbols and Comparison from compare_parts — with a single
frozen Alignment pyclass covering symbol-paired edges, residue
clusters, and unmatched extras on the same shape. Adds cached
qstr / rstr (space-joined comparable forms) so literal-equality
checks and __str__ rendering don't rebuild on every call;
hash/eq key on (symbol, qps, rps) so position survives via
NamePart's (index, form) hash.

Module reshuffle: the old Rust alignment.rs (person-name-order
helper) renamed to ordering.rs to free the filename for the new
Alignment pyclass. Python mirrors the split — the type lives in
rigour.names.compare, ordering.py hosts align_person_name_order,
and the empty alignment.py wrapper is gone.

pair_symbols ships score=1.0 placeholder; the per-category
override (NAME → 0.9 etc.) is matcher policy and lives downstream
in nomenklatura's SYM_SCORES, not in rigour.

Verified: 465 pytest, 185 cargo test, mypy --strict, clippy
(both with and without --features python), cargo fmt — all
clean. arch-name-pipeline.md updated with the new Alignment
section and revised pair_symbols.
